Move Button contents into correct dir
