+++ draft-packages/button/index.ts
@@ -1,2 +1,2 @@
-export { default as Button } from "./KaizenDraft/Button"
-export { default as IconButton } from "./KaizenDraft/IconButton"
+export { default as Button } from "./KaizenDraft/Button/Button"
+export { default as IconButton } from "./KaizenDraft/Button/IconButton"
